move the hook useCursorPaginationForDataTable into a dedicated file
